requirements.txt
gunicorn==18.0
pymongo==2.6.3
wsgiref==0.1.2
git+https://github.com/django-nonrel/django@nonrel-1.5
git+https://github.com/django-nonrel/djangotoolbox
git+https://github.com/django-nonrel/mongodb-engine
django-gravatar2==1.1.3
python-gettext==2.1
Pillow==2.4
python-social-auth
djangorestframework==2.3.13
django-filter==0.7